requirements.txt
python-dotenv
py-cord[voice]
pydub
